[type:feature] plugin enabled adapted namespace
