+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/service/impl/NamespacePluginServiceImpl.java
@@ -49,6 +49,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @Service
@@ -178,7 +179,38 @@ public class NamespacePluginServiceImpl implements NamespacePluginService {
         }
         return StringUtils.EMPTY;
     }
-
+    
+    @Override
+    public String enabled(final String namespaceId, final List<String> pluginIds, final Boolean enabled) {
+        
+        List<NamespacePluginVO> namespacePluginList = namespacePluginRelMapper.selectByNamespaceId(namespaceId);
+        
+        if (CollectionUtils.isEmpty(namespacePluginList)) {
+            return StringUtils.EMPTY;
+        }
+        
+        Map<String, NamespacePluginVO> namespacePluginMap = namespacePluginList
+                .stream()
+                .collect(Collectors.toMap(NamespacePluginVO::getPluginId, Function.identity()));
+        
+        List<NamespacePluginVO> updateList = Lists.newArrayList();
+        
+        for (String pluginId : pluginIds) {
+            NamespacePluginVO namespacePluginVO = namespacePluginMap.get(pluginId);
+            if (Objects.isNull(namespacePluginVO)) {
+                return AdminConstants.SYS_PLUGIN_ID_NOT_EXIST;
+            }
+            namespacePluginVO.setEnabled(enabled);
+            updateList.add(namespacePluginVO);
+        }
+        namespacePluginRelMapper.updateEnableByNamespaceIdAndPluginIdList(namespaceId, pluginIds, enabled);
+        // publish change event.
+        if (CollectionUtils.isNotEmpty(updateList)) {
+            namespacePluginEventPublisher.onEnabled(updateList);
+        }
+        return StringUtils.EMPTY;
+    }
+    
     @Override
     public List<PluginSnapshotVO> activePluginSnapshot() {
         //todo:Not yet  implemented
