+++ shenyu-admin/src/main/resources/mappers/namespace-plugin-rel-sqlmap.xml
@@ -71,6 +71,13 @@
         </foreach>
     </insert>
 
+    <select id="listByNamespaceId" resultMap="BaseResultMap">
+        select
+        <include refid="Base_Column_List"/>
+        from namespace_plugin_rel
+        where namespace_id = #{namespaceId}
+    </select>
+    
     <select id="selectByQuery" parameterType="org.apache.shenyu.admin.model.query.NamespacePluginQuery"
             resultType="org.apache.shenyu.admin.model.vo.NamespacePluginVO">
         SELECT
@@ -265,7 +272,7 @@
         p.plugin_jar AS pluginJar
         FROM namespace_plugin_rel npr
         LEFT JOIN plugin p ON p.id = npr.plugin_id
-        AND npr.namespace_id=#{namespaceId}
+        WHERE npr.namespace_id=#{namespaceId}
         ORDER BY
         sort,
         npr.id ASC
@@ -286,7 +293,7 @@
         p.plugin_jar AS pluginJar
         FROM namespace_plugin_rel npr
         LEFT JOIN plugin p ON p.id = npr.plugin_id
-        AND npr.namespace_id IN
+        WHERE npr.namespace_id IN
         <foreach item="namespaceId" collection="namespaceIds" open="(" separator="," close=")">
             #{namespaceId, jdbcType=VARCHAR}
         </foreach>
@@ -344,6 +351,19 @@
         <foreach collection="idList" item="id" index="index"
                  open="(" separator="," close=")">
             #{id, jdbcType=VARCHAR}
+        </foreach>
+    </update>
+    
+    <update id="updateEnableByNamespaceIdAndPluginIdList">
+        UPDATE namespace_plugin_rel
+        SET enabled = #{enabled, jdbcType=TINYINT}
+        WHERE
+        namespace_id=#{namespaceId}
+        AND
+        plugin_id IN
+        <foreach collection="pluginIds" item="pluginId" index="index"
+                 open="(" separator="," close=")">
+            #{pluginId, jdbcType=VARCHAR}
         </foreach>
     </update>
